wip(content): Floaty Filter and ``panflute`` Filters Consolidation.
Defined a ``BaseFilter`` since I like the pattern.
Updated watcher to watch ``scripts.filters`` by default.
Made ``scripts.filters`` to keep all python code together.
Added links, projects, and education to the resume filter.
