+++ blog/projects/blog/floaty.py
@@ -1,0 +1,4 @@
+from scripts.filters import floaty
+
+if __name__ == "__main__":
+    floaty.filter()
